fully working widget with all profiles including custom

+++ src/dev/ukanth/ufirewall/widget/ToggleWidgetActivity.java
@@ -16,6 +16,7 @@ import android.widget.Toast;
 import dev.ukanth.ufirewall.Api;
 import dev.ukanth.ufirewall.G;
 import dev.ukanth.ufirewall.R;
+import dev.ukanth.ufirewall.RootShell.RootCommand;
 import dev.ukanth.ufirewall.widget.RadialMenuWidget.RadialMenuEntry;
 
 public class ToggleWidgetActivity extends Activity {
@@ -88,8 +89,8 @@ public class ToggleWidgetActivity extends Activity {
 	  
 	  public class Status implements RadialMenuEntry
 	   {
-	      public String getName() { return ""; } 
-		  public String getLabel() { return ""; } 
+	      public String getName() { return G.getActiveProfileName(getApplicationContext()); } 
+		  public String getLabel() { return G.getActiveProfileName(getApplicationContext()); } 
 		  public int getIcon() { return (Api.isEnabled(getApplicationContext()) ?  R.drawable.widget_on :  R.drawable.widget_off); }
 	      public List<RadialMenuEntry> getChildren() { return null; }
 	      public void menuActiviated()
@@ -118,10 +119,42 @@ public class ToggleWidgetActivity extends Activity {
 	      public int getIcon() { return 0; }
 	      private List<RadialMenuEntry> children = new ArrayList<RadialMenuEntry>( Arrays.asList( new DefaultProfile(), new Profile1(), new Profile2(), new Profile3() ) );
 	      public List<RadialMenuEntry> getChildren() { return children; }
+	      public Profiles(){
+	    	  for(String profileName: G.getAdditionalProfiles()) {
+	    		  RadialMenuEntry entry = new GenericProfile(profileName);
+	    		  children.add(entry);
+	    	  }
+	      }
 	      public void menuActiviated()
 	      {
 	      }
 	   }	
+	   
+	   public class GenericProfile implements RadialMenuEntry
+       {
+          public String getName() { return profileName; } 
+          public String getLabel() { return profileName; }
+          public int getIcon() { return 0; }
+          private String profileName;
+          public GenericProfile(String profileName){
+        	  this.profileName = profileName;
+          }
+          public List<RadialMenuEntry> getChildren() { return null; }
+          public void menuActiviated()
+          {
+        	  int pos = G.getProfilePosition(profileName);
+        	  final Message msg = new Message();
+        	  Toast.makeText(getApplicationContext(), profileName + " pressed. + position" + pos, Toast.LENGTH_SHORT).show();
+        	  final Handler toaster = new Handler() {
+      			public void handleMessage(Message msg) {
+      				if (msg.arg1 != 0)Toast.makeText(getApplicationContext(), msg.arg1, Toast.LENGTH_SHORT).show();
+      			}
+      		};
+      		final Context context = getApplicationContext();
+      		G.setProfile(true, pos);
+			applyProfileRules(context,msg,toaster);
+          }
+       }    
 	   
 	   public class DefaultProfile implements RadialMenuEntry
        {
@@ -186,49 +219,45 @@ public class ToggleWidgetActivity extends Activity {
 			public void run() {
 				Looper.prepare();
 				final Message msg = new Message();
-				switch(i){
-				case 1: 
-					if(applyRules(context,msg,toaster)){
-						Api.setEnabled(context, true, false);
-					}
-					break;
-				case 2:
-					//validation, check for password
-					if(oldPwd.length() == 0 && newPwd.length() == 0){
-						if (Api.purgeIptables(context, false)) {
-							msg.arg1 = R.string.toast_disabled;
-							toaster.sendMessage(msg);
-							Api.setEnabled(context, false, false);
+				if( i < 7) {
+					switch(i){
+					case 1: 
+						if(applyRules(context,msg,toaster)){
+							Api.setEnabled(context, true, false);
+						}
+						break;
+					case 2:
+						//validation, check for password
+						if(oldPwd.length() == 0 && newPwd.length() == 0){
+							if (Api.purgeIptables(context, false)) {
+								msg.arg1 = R.string.toast_disabled;
+								toaster.sendMessage(msg);
+								Api.setEnabled(context, false, false);
+							} else {
+								msg.arg1 = R.string.toast_error_disabling;
+								toaster.sendMessage(msg);
+							}
 						} else {
-							msg.arg1 = R.string.toast_error_disabling;
+							msg.arg1 = R.string.widget_disable_fail;
 							toaster.sendMessage(msg);
 						}
-					} else {
-						msg.arg1 = R.string.widget_disable_fail;
-						toaster.sendMessage(msg);
+						break;
+					case 3:
+						G.setProfile(G.enableMultiProfile(), 0);
+						break;
+					case 4:
+						G.setProfile(true, 1);
+						break;
+					case 5:
+						G.setProfile(true, 2);
+						break;
+					case 6:
+						G.setProfile(true, 3);
+						break;
 					}
-					
-					break;
-				case 3:
-					G.setProfile(G.enableMultiProfile(), 0);
-					if(applyProfileRules(context,msg,toaster)) {
+					if(i > 2) {
+						applyProfileRules(context,msg,toaster);
 					}
-					break;
-				case 4:
-					G.setProfile(true, 1);
-					if(applyProfileRules(context,msg,toaster)){
-					}
-					break;
-				case 5:
-					G.setProfile(true, 2);
-					if(applyProfileRules(context,msg,toaster)){
-					}
-					break;
-				case 6:
-					G.setProfile(true, 3);
-					if(applyProfileRules(context,msg,toaster)){
-					}
-					break;
 				}
 			}
 		}.start();
@@ -249,17 +278,33 @@ public class ToggleWidgetActivity extends Activity {
 		return success;
 	}
 	
-	private boolean applyProfileRules(Context context,Message msg, Handler toaster) {
+	private boolean applyProfileRules(final Context context,final Message msg, final Handler toaster) {
 		boolean success = false;
-		if (Api.applySavedIptablesRules(context, false)) {
+/*		if (Api.applySavedIptablesRules(context, false)) {
 			msg.arg1 = R.string.rules_applied;
 			toaster.sendMessage(msg);
 			success = true;
 		} else {
 			msg.arg1 = R.string.error_apply;
 			toaster.sendMessage(msg);
-		}
-		return success;
+		}*/
+		
+		boolean ret = Api.fastApply(context, new RootCommand()
+		.setFailureToast(R.string.error_apply)
+		.setCallback(new RootCommand.Callback() {
+			@Override
+			public void cbFunc(RootCommand state) {
+				if (state.exitCode == 0) {
+					msg.arg1 = R.string.rules_applied;
+					toaster.sendMessage(msg);
+				} else {
+					// error details are already in logcat
+					msg.arg1 = R.string.error_apply;
+					toaster.sendMessage(msg);
+				}
+			}
+		}));
+		return ret;
 	}
 }
